@@ -1,42 +1,28 @@
 {
-  "name": "my-ttrpg-generagots",
-  "version": "1.0.1",
+  "name": "everlight-quest",
+  "private": true,
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development vite",
+    "dev": "vite",
     "build": "vite build",
-    "test": "jest",
-    "test:watch": "jest --watchAll",
-    "serve": "vite preview"
+    "preview": "vite preview",
+    "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
-    "@netlify/functions": "^1.6.0",
-    "@sveltejs/vite-plugin-svelte": "^1.0.1",
-    "@tailwindcss/typography": "^0.5.9",
-    "@testing-library/svelte": "^3.0.3",
-    "@types/jest": "^26.0.21",
-    "autoprefixer": "^10.2.5",
-    "babel-jest": "^26.6.3",
-    "cross-env": "^7.0.3",
-    "cssnano": "^5.1.12",
-    "jest": "^26.6.3",
-    "netlify-cli": "^15.1.1",
-    "postcss": "^8.4.6",
-    "postcss-load-config": "^4.0.1",
-    "svelte": "^3.35.0",
-    "svelte-jester": "^1.3.1",
-    "svelte-preprocess": "^4.10.7",
-    "tailwindcss": "^3.1.6",
-    "ts-jest": "^26.5.4",
-    "typescript": "^4.9.5",
-    "vite": "^3.0.0",
-    "vite-plugin-pwa": "^0.15.1"
+    "@sveltejs/vite-plugin-svelte": "^3.0.1",
+    "@tsconfig/svelte": "^5.0.2",
+    "autoprefixer": "^10.4.16",
+    "postcss": "^8.4.32",
+    "svelte": "^4.2.8",
+    "svelte-check": "^3.6.2",
+    "svelte-routing": "^2.10.0",
+    "tailwindcss": "^3.3.6",
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.8"
   },
   "dependencies": {
-    "@dice-roller/rpg-dice-roller": "^5.2.1",
-    "@types/node": "^14.18.47",
-    "canvas": "^2.11.2",
-    "markdown-it": "^13.0.1",
-    "svelte-routing": "^1.6.0",
-    "turndown": "^7.1.2"
+    "@dice-roller/rpg-dice-roller": "^5.5.0"
   }
 }
