@@ -1,8 +1,7 @@
-const sveltePreprocess = require("svelte-preprocess");
-module.exports = {
-	preprocess: [
-		sveltePreprocess({
-			postcss: true
-		}),
-	],
-};
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte'
+
+export default {
+  // Consult https://svelte.dev/docs#compile-time-svelte-preprocess
+  // for more information about preprocessors
+  preprocess: vitePreprocess(),
+}
