@@ -1,21 +1,12 @@
-module.exports = {
-  mode: 'jit',
-  content: ['./src/**/*.{html,js,svelte,ts}'],
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,svelte,html}",
+  ],
   theme: {
-    extend: {
-      fontFamily: [
-        '-apple-system',
-        'BlinkMacSystemFont',
-        'Segoe UI',
-        'Roboto',
-        'Oxygen',
-        'Ubuntu',
-        'Cantarell',
-        'Open Sans',
-        'Helvetica Neue',
-        'sans-serif',
-      ],
-    },
+    extend: {},
   },
-  plugins: [require('@tailwindcss/typography')],
-};
+  plugins: [],
+}
+
