@@ -1,9 +1,8 @@
-import './hmr';
-import './main.css';
+import './assets/style.css';
 import App from './App.svelte'
 
 const app = new App({
-  target: document.body,
+  target: document.getElementById('app') as Element,
 })
 
 export default app
